update v1.3.5 hobby,tutoring create,update

+++ api/app/Http/Controllers/TutoringController.php
@@ -29,18 +29,18 @@ class TutoringController extends Controller
             $tutoringDb = GroupModel::where('type', 'tutoring')
                 ->where('status', 1)
                 ->orderBy('updated_at', 'DESC')
-                ->with('tutoring')
-                ->with('bookmark')
-                ->with(['tutoring.imageOrFile'])
-                ->with(['tutoring.leaderGroup'])
-                ->with(['tutoring.faculty'])
-                ->with(['tutoring.major'])
-                ->with(['tutoring.department'])
-                ->with('member')
-                ->with('request')
-                ->with('groupDay')
-                ->with('groupTag')
-                ->get();
+                // ->with('tutoring')
+                // ->with('bookmark')
+                // ->with(['tutoring.imageOrFile'])
+                // ->with(['tutoring.leaderGroup'])
+                // ->with(['tutoring.faculty'])
+                // ->with(['tutoring.major'])
+                // ->with(['tutoring.department'])
+                // ->with('member')
+                // ->with('request')
+                // ->with('groupDay')
+                // ->with('groupTag')
+                ->paginate(8);
 
             if (!$tutoringDb) {
                 return response()->json([
@@ -54,14 +54,14 @@ class TutoringController extends Controller
                 return response()->json([
                     // 'prevPageUrl' => $tutoringDb->previousPageUrl(),
                     'status' => 'ok',
-                    'message' => 'fetch all hobby-group success.',
+                    'message' => 'fetch all tutoring group success.',
                     'listItem' => $data,
                     // 'nextPageUrl' => $tutoringDb->nextPageUrl()
                 ], 200);
             } else {
                 return response()->json([
                     'status' => 'failed',
-                    'message' => 'failed to fetch all hobby group.',
+                    'message' => 'failed to fetch tutoring group.',
                 ], 500);
             };
         }
@@ -85,11 +85,11 @@ class TutoringController extends Controller
 
         try {
             $uID = auth()->user()->id;
-            $path = public_path('uploaded/hobbyImage/');
+            $path = public_path('uploaded\\hobbyImage\\');
             if ($request->file('image') != null) {
                 $file = $request->file('image');
                 $extension = $file->getClientOriginalExtension();
-                $filename = 'hobby-' . date('YmdHi') . '.' . $extension;
+                $filename = 'tutoring-' . date('YmdHi') . '.' . $extension;
                 $file->move($path, $filename);
                 $imageOrFileDb = imageOrFileModel::create([
                     'name' => $filename
@@ -119,42 +119,16 @@ class TutoringController extends Controller
                 'name' => $request->input('activityName'),
                 'memberMax' => $request->input('memberMax') ?? null,
                 'location' =>  $request->input('location'),
-                'detail' =>  $request->input('detail'),
-                'startTime' =>  $request->input('actTime') ?? date_format(now(), "H:i:s"),
-                'endTime' => $request->input('actTime') ?? date("H:i:s", mktime(0, 0, 0)),
-                'date' => $request->input('date') ?? date("Y-m-d"),
+                'detail' =>  $request->input('detail') ?? null,
+                'startTime' =>  date("H:i:s", strtotime($request->input('startTime'))) ?? date_format(now(), "H:i:s"),
+                'endTime' => date("H:i:s", strtotime($request->input('endTime'))) ?? date("H:i:s", mktime(0, 0, 0)),
+                'date' => date("Y-m-d", strtotime($request->input('date'))) ?? date("Y-m-d"),
                 'leader' => $uID,
                 'createdBy' => $uID,
                 'created_at' => now(),
                 'updated_at' => now(),
             ]);
             //-----------
-
-
-            //----------------------groupday part
-            if ($request->has('weekDate')) {
-                $days = explode(',', $request->input('weekDate'));
-                //loop through array of days input and save to Db
-                foreach ($days as $day) {
-                    //if array is empty then set day to today of weekday (it will has only one array)
-                    if ($day == '' || $day == null) {
-                        $day = (int)date('w');
-                    }
-                    GroupDayModel::create([
-                        'dayID' => (int)$day,
-                        'groupID' => $groupDb->id,
-                    ]);
-                }
-            }
-            //using today of weekday if days input is empty
-            else {
-                $day = date('w');
-                GroupDayModel::create([
-                    'dayID' => (int)$day,
-                    'groupID' => $groupDb->id,
-                ]);
-            }
-            //----------------------
 
             //----------------------tag part
             $tags = explode(',', $request->input('tag'));
@@ -204,11 +178,18 @@ class TutoringController extends Controller
                 'trace' => $e->getTraceAsString()
             ]);
             //delete group with all its related model when failed
-            GroupTagModel::where('groupID', $groupDb->id)->delete();
-            GroupDayModel::where('groupID', $groupDb->id)->delete();
-            MemberModel::where('groupID', $groupDb->id)->delete();
-            TutoringModel::where('id', $tID)->delete();
-            GroupModel::where('groupID', $tID)->delete();
+            if (GroupTagModel::where('groupID', $groupDb->id)->first()) {
+                GroupTagModel::where('groupID', $groupDb->id)->delete();
+            }
+            if (MemberModel::where('groupID', $groupDb->id)->first()) {
+                MemberModel::where('groupID', $groupDb->id)->delete();
+            }
+            if (TutoringModel::where('id', $tID)->first()) {
+                TutoringModel::where('id', $tID)->delete();
+            }
+            if (GroupModel::where('groupID', $tID)->first()) {
+                GroupModel::where('groupID', $tID)->delete();
+            }
             return response()->json([
                 'status' => 'failed',
                 'message' => 'failed to create tutoring group.'
@@ -221,8 +202,6 @@ class TutoringController extends Controller
         //done (waitng for checking) **noti to everymem
         $uID = auth()->user()->id;
 
-        //ปิด validate หากต้องการเทสเอง
-        //ถ้าเปิด ต้องให้ frontend ส่งค่าเก่าติดมาด้วย
         $validationRules = TutoringModel::$validator[0];
         $validationMessages = TutoringModel::$validator[1];
 
@@ -235,7 +214,7 @@ class TutoringController extends Controller
             ], 400);
         }
 
-        //เรียกใช้ relation
+        //----- เรียกใช้ relation
         $groupDb = GroupModel::where('groupID', $tID)
             ->where([['type', 'tutoring'], ['status', 1]])
             ->with(['tutoring', 'tutoring.imageOrFile', 'groupDay', 'groupTag', 'member'])
@@ -248,112 +227,95 @@ class TutoringController extends Controller
                 'message' => 'Group not found.'
             ], 404);
         }
+        //---------------
 
-        //ถ้ามีการใส่รูป
+        //----------------- ถ้ามีการใส่รูป
+        $path = public_path('uploaded\\hobbyImage\\');
+        //--- ชื่อไฟล์ที่ใช้ seed
+        $defaultFiles = [];
+        foreach (imageOrFileModel::$groupImageStatic as $file) {
+            array_push($defaultFiles, $file['name']);
+        }
+        //---
         if (!empty($request->file('image'))) {
-            //เมื่อไม่ใช่รูปเก่า หรือรูป default (มีการอัพรูปใหม่)
-            if ($groupDb->tutoring->imageOrFile->name !== 'group-default.jpg' && $groupDb->tutoring->imageOrFile->name !== $request->file('image')) {
-                //path รูปภาพ
-                $path = public_path('uploaded/hobbyImage/');
-                File::delete($path . $groupDb->tutoring->imageOrFile->name); //ลบรูปเก่าทิ้ง
-                imageOrFileModel::where('id',$groupDb->tutoring->imageOrFile->id)->delete(); // ลบ data on db
-
-                $file = $request->file('image');
-                $extension = $file->getClientOriginalExtension();
-                $filename = 'tutoring-' . now()->format('YmdHis') . '.' . $extension;
-                $move = $file->move($path, $filename);
-
-                if (!$move) {
-                    return response()->json([
-                        'status' => 'failed',
-                        'message' => 'Can not upload image.'
-                    ], 500);
-                } else {
-                    $imageOrFileDb = imageOrFileModel::create([
-                        'name' => $filename
-                    ]);
+            $file = $request->file('image');
+            if (imageOrFileModel::where('id', $groupDb->tutoring->imageOrFile->id)->first() && !in_array(strval($groupDb->tutoring->imageOrFile->name), $defaultFiles)) {
+                imageOrFileModel::where('id', $groupDb->tutoring->imageOrFile->id)->delete(); // ลบชื่อไฟล์บน database
+                if (File::exists($path . $groupDb->tutoring->imageOrFile->name)) {
+                    File::delete($path . $groupDb->tutoring->imageOrFile->name); //ลบไฟล์รูปเก่าทิ้ง
                 }
             }
-        }
-        else {
-            $imageOrFileDb = imageOrFileModel::where('id',$groupDb->tutoring->imageOrFile->id)->first();
-        }
-
-        //แตก array นั้นลบ tag เก่าทั้งหมดออก แต่จะไม่ลบตัว static
-        //จากนั้นสร้างใหม่ใน GroupTagModel โดยเก็บ tag ใหม่ใน TagModel
-        //----------------------tag part
-
-        //-----------------------
-
-        $newTags = explode(',', $request->input('tag'));
-        $deleteOldTag = GroupTagModel::where('groupID', $groupDb->id)->where('type', 'tutoring')->delete();
-
-        if ($request->input('tag') == '' || $request->input('tag') == null) {
-            $newTags = ['tutoring'];
-        }
-
-        foreach ($newTags as $tag) {
-            $tagInTagModel = TagModel::where('name', $tag)->first();
-            if (empty($tagInTagModel) && $tag != '') {
-                $newTag = TagModel::create([
-                    'name' => $tag
-                ]);
-                $tagGroup = $newTag->id;
+            $extension = $file->getClientOriginalExtension();
+            $filename = 'tutoring-' . now()->format('YmdHis') . '.' . $extension;
+            $move = $file->move($path, $filename);
+            if (!$move) {
+                return response()->json([
+                    'status' => 'failed',
+                    'message' => 'Can not upload image.'
+                ], 500);
             } else {
-                $tagGroup = $tagInTagModel->id;
-            }
-
-            GroupTagModel::create([
-                'groupID' => $groupDb->id,
-                'tagID' => $tagGroup,
-                'type' => 'tutoring'
-            ]);
-        };
-        // -----------------------------
-
-        // แตก array เหมือน tag และลบอันเก่าทิ้งเพื่อสร้างใหม่ เผื่อกรณีเพิ่มวัน
-        $deleteOldDay = GroupDayModel::where('groupID', $groupDb->id)->delete();
-
-        if ($request->has('weekDate')) {
-            $days = explode(',', $request->input('weekDate'));
-            //loop through array of days input and save to Db
-            foreach ($days as $day) {
-                //if array is empty then set day to today of weekday (it will has only one array)
-                if ($day == '' || $day == null) {
-                    $day = (int)date('w');
-                }
-                GroupDayModel::create([
-                    'dayID' => (int)$day,
-                    'groupID' => $groupDb->id,
+                $imageOrFileDb = imageOrFileModel::create([
+                    'name' => $filename
                 ]);
             }
-        } else {
-            $day = (int)date('w');
-            GroupDayModel::create([
-                'dayID' => (int)$day,
-                'groupID' => $groupDb->id,
-            ]);
+        } else if (empty($request->file('image')) && !$request->has('deleteimage')) {
+            $imageOrFileDb = imageOrFileModel::where('id', $groupDb->tutoring->imageOrFile->id)->first();
+        } else if ($request->has('deleteimage')) {
+            if (imageOrFileModel::where('id', $groupDb->tutoring->imageOrFile->id)->first() && !in_array(strval($groupDb->tutoring->imageOrFile->name), $defaultFiles,true)) {
+                imageOrFileModel::where('id', $groupDb->tutoring->imageOrFile->id)->delete(); // ลบชื่อไฟล์บน database
+                if (File::exists($path . $groupDb->tutoring->imageOrFile->name)) {
+                    File::delete($path . $groupDb->tutoring->imageOrFile->name); //ลบไฟล์รูปเก่าทิ้ง
+                }
+                $imageOrFileDb = imageOrFileModel::where('name', 'group-default.jpg')->first();
+            }
         }
+        //----------------------
 
-        //------------------------------ 
+        //----------------- tag part
+        if ($request->input('tag')) {
+            $deleteOldTag = GroupTagModel::where('groupID', $groupDb->id)->where('type', 'tutoring')->delete();
+            $newTags = explode(',', $request->input('tag'));
+
+            foreach ($newTags as $tag) {
+                if ($tag != '' || $tag != null) {
+                    $tag = ['tutoring'];
+                }
+                $querytag = TagModel::where('name', $tag)->first();
+                if (empty($querytag)) {
+                    $newTag = TagModel::create([
+                        'name' => $tag
+                    ]);
+                    $tagGroup = $newTag->id;
+                } else {
+                    $tagGroup = $querytag->id;
+                }
+
+                GroupTagModel::create([
+                    'groupID' => $groupDb->id,
+                    'tagID' => $tagGroup,
+                    'type' => 'tutoring'
+                ]);
+            };
+        }
+        // -----------------------------
 
         $data = [
             'facultyID' => $request->input('facultyID') ?? $groupDb->tutoring->facultyID,
             'majorID' => $request->input('majorID') ?? $groupDb->tutoring->majorID,
             'departmentID' => $request->input('departmentID') ?? $groupDb->tutoring->departmentID,
-            'imageOrFileID' => $imageOrFileDb->id ?? $groupDb->tutoring->imageOrFile->id ?? imageOrFileModel::where('name','group-default.jpg')->first()->id,
+            'imageOrFileID' => $imageOrFileDb->id ?? $groupDb->tutoring->imageOrFile->id,
             'name' => $request->input('activityName') ?? $groupDb->tutoring->activityName,
             'memberMax' => $request->input('memberMax') ?? $groupDb->tutoring->memberMax,
             'location' => $request->input('location') ?? $groupDb->tutoring->location,
             'detail' => $request->input('detail') ?? $groupDb->tutoring->detail ?? null,
-            'startTime' => $request->input('startTime') ?? $groupDb->tutoring->startTime,
-            'endTime' => $request->input('endTime') ?? $groupDb->tutoring->endTime,
-            'date' => $request->input('date') ?? date("Y-m-d"),
+            'startTime' => date("H:i:s", strtotime($request->input('startTime'))) ?? $groupDb->tutoring->startTime,
+            'endTime' => date("H:i:s", strtotime($request->input('endTime'))) ?? $groupDb->tutoring->endTime,
+            'date' => date("Y-m-d", strtotime($request->input('date'))) ?? date("Y-m-d"),
             'leader' => $uID,
             'updated_at' => now()
         ];
 
-        // อัพเดตข้อมูลทั่วไปของ hobby และส่งแจ้งเตือนอัพเดต
+        // อัพเดตข้อมูลทั่วไปของ tutoring และส่งแจ้งเตือนอัพเดต
         if (TutoringModel::where('id', $tID)->update($data)) {
             foreach ($groupDb->member as $member) {
                 if ($member->id != $uID) {
@@ -643,7 +605,6 @@ class TutoringController extends Controller
                 MemberModel::where('groupID', $groupDb->id)->delete() && GroupTagModel::where('groupID', $groupDb->id)->delete()
                 && GroupDayModel::where('groupID', $groupDb->id)->delete() && GroupModel::where('groupID', $tID)->delete()
                 && TutoringModel::where('id', $tID)->delete()
-                
             ) {
                 foreach ($groupDb->member as $member) {
                     NotifyModel::create([
